+++ services/autorust/codegen/src/codegen_operations/web_operation_gen.rs
@@ -10,7 +10,7 @@ use crate::{
     WebOperation,
 };
 
-use super::{create_function_name, API_VERSION, X_MS_VERSION};
+use super::{API_VERSION, X_MS_VERSION};
 
 pub struct WebOperationGen(pub WebOperation);
 
@@ -128,4 +128,57 @@ impl WebOperationGen {
 #[derive(Clone)]
 pub struct Pageable {
     pub next_link_name: Option<String>,
+}
+
+/// Creating a function name from the path and verb when an operationId is not specified.
+/// All azure-rest-api-specs operations should have an operationId.
+fn create_function_name(verb: &WebVerb, path: &str) -> String {
+    let mut path = path.split('/').filter(|&x| !x.is_empty()).collect::<Vec<_>>();
+    path.insert(0, verb.as_str());
+    path.join("_")
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_create_function_name() {
+        assert_eq!(create_function_name(&WebVerb::Get, "/pets"), "get_pets");
+    }
+
+    #[test]
+    fn test_function_name_from_operation_id() {
+        let operation = WebOperationGen(WebOperation {
+            id: Some("PrivateClouds_CreateOrUpdate".to_owned()),
+            path: "/horse".to_owned(),
+            verb: WebVerb::Get,
+            ..Default::default()
+        });
+        assert_eq!(Some("private_clouds".to_owned()), operation.rust_module_name());
+        assert_eq!("create_or_update", operation.rust_function_name());
+    }
+
+    #[test]
+    fn test_function_name_from_verb_and_path() {
+        let operation = WebOperationGen(WebOperation {
+            path: "/horse".to_owned(),
+            verb: WebVerb::Get,
+            ..Default::default()
+        });
+        assert_eq!(None, operation.rust_module_name());
+        assert_eq!("get_horse", operation.rust_function_name());
+    }
+
+    #[test]
+    fn test_function_name_with_no_module_name() {
+        let operation = WebOperationGen(WebOperation {
+            id: Some("PerformConnectivityCheck".to_owned()),
+            path: "/horse".to_owned(),
+            verb: WebVerb::Put,
+            ..Default::default()
+        });
+        assert_eq!(None, operation.rust_module_name());
+        assert_eq!("perform_connectivity_check", operation.rust_function_name());
+    }
 }
